+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.6.2
 * Mercator 1.0.4
+* Fix delimiting of AWS regions
 
 ### 1.6.1
 * Mercator 1.0.2